requirements-dev.txt
ruff==0.4.4
mypy==1.10.0
bandit==1.7.8
pytest==8.2.0
pytest-flask==1.3.0
pytest-cov==5.0.0
pytest-mock==3.14.0
pre-commit==3.7.1
requests-mock
